# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.2.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
}

export type VillaStatus = typeof VillaStatus[keyof typeof VillaStatus];


export const VillaStatus = {
  draft: 'draft',
  published: 'published',
  sold: 'sold',
  archived: 'archived',
  inactive: 'inactive',
} as const;

export interface Villa {
  id: number;
  villa_code: string;
  city?: string | null;
  area_type?: string | null;
  price?: number | null;
  land_size?: number | null;
  building_size?: number | null;
  bedrooms?: number | null;
  master_bedrooms?: number | null;
  is_townhouse: number;
  has_pool: number;
  has_jacuzzi: number;
  has_roof_garden: number;
  has_parking: number;
  has_storage: number;
  document_type?: string | null;
  description?: string | null;
  latitude?: number | null;
  longitude?: number | null;
  photos?: string | null;
  video?: string | null;
  status: VillaStatus;
  created_at: string;
  updated_at: string;
}

export interface CreateVillaRequest {
  /** Optional MV code (e.g. "MV-1042"). When provided the server uses it as-is and skips auto-generation. When omitted the server assigns the next available code automatically. */
  villa_code?: string | null;
  city?: string | null;
  area_type?: string | null;
  price?: number | null;
  land_size?: number | null;
  building_size?: number | null;
  bedrooms?: number | null;
  master_bedrooms?: number | null;
  is_townhouse?: number;
  has_pool?: number;
  has_jacuzzi?: number;
  has_roof_garden?: number;
  has_parking?: number;
  has_storage?: number;
  document_type?: string | null;
  description?: string | null;
  latitude?: number | null;
  longitude?: number | null;
  photos?: string | null;
  video?: string | null;
  status?: VillaStatus;
}

export interface UpdateVillaRequest {
  city?: string | null;
  area_type?: string | null;
  price?: number | null;
  land_size?: number | null;
  building_size?: number | null;
  bedrooms?: number | null;
  master_bedrooms?: number | null;
  is_townhouse?: number | null;
  has_pool?: number | null;
  has_jacuzzi?: number | null;
  has_roof_garden?: number | null;
  has_parking?: number | null;
  has_storage?: number | null;
  document_type?: string | null;
  description?: string | null;
  latitude?: number | null;
  longitude?: number | null;
  photos?: string | null;
  video?: string | null;
  status?: VillaStatus;
}

export interface UpdateVillaStatusRequest {
  status: VillaStatus;
}

export type VillaStatsByCityItem = {
  city: string;
  count: number;
};

export type VillaStatsByPriceTierItem = {
  tier: string;
  count: number;
};

export interface VillaStats {
  total: number;
  published: number;
  draft: number;
  sold: number;
  archived: number;
  by_city: VillaStatsByCityItem[];
  by_price_tier: VillaStatsByPriceTierItem[];
}

export interface VillasPage {
  data: Villa[];
  total: number;
  page: number;
  page_size: number;
}

export interface VisitRequest {
  id: number;
  villa_code: string;
  user_id: number;
  name: string;
  phone: string;
  area_type?: string | null;
  request_type: string;
  status: string;
  created_at: string;
  price?: number | null;
  villa_city?: string | null;
}

export interface RequestsPage {
  data: VisitRequest[];
  total: number;
  page: number;
  page_size: number;
}

export interface RequestStats {
  total: number;
  pending: number;
  contacted: number;
  visit_count: number;
  consultation_count: number;
}

export type ListVillasParams = {
status?: VillaStatus;
city?: string;
area_type?: string;
page?: number;
page_size?: number;
};

export type HardDeleteVilla200 = {
  deleted?: boolean;
  id?: number;
};

export type ListRequestsParams = {
status?: ListRequestsStatus;
request_type?: ListRequestsRequestType;
page?: number;
page_size?: number;
};

export type ListRequestsStatus = typeof ListRequestsStatus[keyof typeof ListRequestsStatus];


export const ListRequestsStatus = {
  pending: 'pending',
  contacted: 'contacted',
} as const;

export type ListRequestsRequestType = typeof ListRequestsRequestType[keyof typeof ListRequestsRequestType];


export const ListRequestsRequestType = {
  visit: 'visit',
  consultation: 'consultation',
} as const;