# lib/api-zod/src/generated/types/villaStatus.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.2.0
 */

export type VillaStatus = typeof VillaStatus[keyof typeof VillaStatus];


export const VillaStatus = {
  draft: 'draft',
  published: 'published',
  sold: 'sold',
  archived: 'archived',
  inactive: 'inactive',
} as const;